Refine ID length implementation and tests
- Updated `recommended_id_length` to accept `usize`.
- Cleaned up `id_length_breakpoints.rs` tests:
  - Removed debug comments.
  - Added descriptive comments for test tuples.
  - Renamed variable `k` to `population`.
- Addresses code review feedback regarding types and test readability.

+++ crates/pebble/src/commands/add.rs
@@ -12,7 +12,7 @@ pub fn run(config: &Config, title: String, description: Option<String>) -> Resul
     let existing_ids: std::collections::HashSet<&str> =
         existing_issues.iter().map(|i| i.id.as_str()).collect();
 
-    let suffix_length = pebble::recommended_id_length(existing_issues.len() as u64);
+    let suffix_length = pebble::recommended_id_length(existing_issues.len());
 
     let mut id;
     loop {
